Abstract the in memory store
