+++ src/TrimDB.Core/InMemory/MemoryIterator.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace TrimDB.Core.InMemory
+{
+    public interface IMemoryItem
+    {
+        ReadOnlySpan<byte> Key { get; }
+        ReadOnlySpan<byte> Value { get; }
+        bool IsDeleted { get; }
+    }
+}
