+++ src/TrimDB.Core/Storage/IOScheduler.cs
@@ -6,22 +6,22 @@ using System.Text;
 using System.Threading;
 using System.Threading.Channels;
 using System.Threading.Tasks;
-using TrimDB.Core.SkipList;
+using TrimDB.Core.InMemory;
 
 namespace TrimDB.Core.Storage
 {
     public class IOScheduler : IAsyncDisposable
     {
-        private Channel<SkipList.SkipList> _channel;
-        private CancellationTokenSource _token = new CancellationTokenSource();
-        private Task _writerTask;
-        private UnsortedStorageLayer _storageLayer;
+        private readonly Channel<MemoryTable> _channel;
+        private readonly CancellationTokenSource _token = new CancellationTokenSource();
+        private readonly Task _writerTask;
+        private readonly UnsortedStorageLayer _storageLayer;
 
         public IOScheduler(int maxSkiplistBacklog, UnsortedStorageLayer storageLayer)
         {
             _storageLayer = storageLayer;
-            _channel = Channel.CreateBounded<SkipList.SkipList>(new BoundedChannelOptions(maxSkiplistBacklog));
-            _writerTask = WriteSkipLists();
+            _channel = Channel.CreateBounded<MemoryTable>(new BoundedChannelOptions(maxSkiplistBacklog));
+            _writerTask = WriteInMemoryTable();
         }
 
         public async ValueTask DisposeAsync()
@@ -31,28 +31,26 @@ namespace TrimDB.Core.Storage
             await _writerTask;
         }
 
-        public ValueTask ScheduleSkipListSave(SkipList.SkipList skipList) => _channel.Writer.WriteAsync(skipList);
+        public ValueTask ScheduleSave(MemoryTable memoryTable) => _channel.Writer.WriteAsync(memoryTable);
 
-        private async Task WriteSkipLists()
+        private async Task WriteInMemoryTable()
         {
             await foreach (var sl in _channel.Reader.ReadAllAsync())
             {
                 try
                 {
-
-
                     var nextFilename = _storageLayer.GetNextFileName();
                     var fileWriter = new TableFileWriter(nextFilename);
-                    await fileWriter.SaveSkipList(sl);
+                    await fileWriter.SaveMemoryTable(sl);
                     var tableFile = new TableFile(fileWriter.FileName);
                     await tableFile.LoadAsync();
                     _storageLayer.AddTableFile(new TableFile(fileWriter.FileName));
                 }
-                catch(Exception ex)
+                catch (Exception ex)
                 {
                     Console.WriteLine("Error " + ex);
                 }
-                //TODO : Remove the skiplist from the database
+                //TODO : Remove the memory table from the database
 
             }
         }
