+++ src/TrimDB.Core/TrimDB.Core.csproj
@@ -1,4 +1,4 @@
-<Project Sdk="Microsoft.NET.Sdk" >
+<Project Sdk="Microsoft.NET.Sdk">
   <ItemGroup>
     <PackageReference Include="System.IO.Pipelines" />
   </ItemGroup>
