+++ tests/TrimDB.Core.Facts/DatabaseFacts.cs
@@ -2,7 +2,8 @@
 using System.Collections.Generic;
 using System.Text;
 using System.Threading.Tasks;
-using TrimDB.Core.SkipList;
+using TrimDB.Core.InMemory.SkipList32;
+using TrimDB.Core.InMemory.SkipList64;
 using Xunit;
 
 namespace TrimDB.Core.Facts
@@ -13,7 +14,7 @@ namespace TrimDB.Core.Facts
         public async Task TestSkipListOverflow()
         {
             var loadedWords = await System.IO.File.ReadAllLinesAsync("words.txt");
-            var db = new TrimDatabase(() => new SkipList.SkipList(new NativeAllocator(4096 * 1024, 25)), 2, "c:\\code\\trimdb\\Database");
+            var db = new TrimDatabase(() => new SkipList32(new NativeAllocator32(4096 * 1024, 25)), 2, "c:\\code\\trimdb\\Database");
 
             foreach (var word in loadedWords)
             {
